pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smallclawlm"
version = "0.3.0"
description = "Zero-token AI agent powered by Google NotebookLM"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [{name = "aissac"}]
keywords = ["notebooklm", "agent", "smolagents", "zero-token"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "smolagents>=0.2",
    "notebooklm-py>=0.3",
    "httpx>=0.27",
    "click>=8.0",
    "rich>=13.0",
]

[project.scripts]
smallclaw = "smallclawlm.cli:cli"

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]
